+++ lib/src/main/kotlin/dev/thecodebuffet/zcash/zip321/model/NonNegativeAmount.kt
@@ -7,7 +7,7 @@ import java.math.RoundingMode
  *
  * @property value The decimal value of the ZEC amount.
  */
-class Amount {
+class NonNegativeAmount {
     private val value: BigDecimal
 
     /**
@@ -70,8 +70,8 @@ class Amount {
          * non-negative non-zero ZEC decimal amount.
          */
         @Throws(AmountError::class)
-        fun create(value: BigDecimal): Amount {
-            return Amount(value.round())
+        fun create(value: BigDecimal): NonNegativeAmount {
+            return NonNegativeAmount(value.round())
         }
 
         /**
@@ -83,7 +83,7 @@ class Amount {
          * amount constraints.
          */
         @Throws(AmountError::class)
-        fun createFromString(string: String): Amount {
+        fun createFromString(string: String): NonNegativeAmount {
             return create(decimalFromString(string))
         }
 
@@ -114,5 +114,20 @@ class Amount {
      */
     override fun toString(): String {
         return value.setScale(maxFractionalDecimalDigits, RoundingMode.HALF_EVEN).stripTrailingZeros().toPlainString()
+    }
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+
+        other as NonNegativeAmount
+
+        // NOTE: comparing with == operator provides false negatives.
+        // apparently this is a JDK issue.
+        return this.value.compareTo(other.value) == 0
+    }
+
+    override fun hashCode(): Int {
+        return 31 * this.value.hashCode()
     }
 }