+++ lib/src/main/kotlin/dev/thecodebuffet/zcash/zip321/parser/AddressParser.kt
@@ -1,0 +1,10 @@
+package dev.thecodebuffet.zcash.zip321.parser
+
+import com.copperleaf.kudzu.parser.text.BaseTextParser
+
+class AddressTextParser : BaseTextParser(
+    isValidChar = { _, char -> char.isLetterOrDigit() },
+    isValidText = { it.isNotEmpty() },
+    allowEmptyInput = false,
+    invalidTextErrorMessage = { "Expected bech32 or Base58 text, got '$it'" }
+)