+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## Unreleased
+
+### Added 
+- ZIP321 enum now has `ParserResult`
+```
+sealed class ParserResult {
+        data class SingleAddress(val singleRecipient: RecipientAddress): ParserResult()
+
+        data class Request(val paymentRequest: PaymentRequest): ParserResult()
+    }
+```
+
+- `fun request(uriString: String, validatingRecipients: ((String) -> Boolean)?): ParserResult`
+- `MemoBytes` now supports `fun fromBase64URL(string: String): MemoBytes`
+
+### modified
+- `Amount` was changed to `NonNegativeAmount`
+
 
 ## [0.0.1] - 2023-11-27